+++ packages/snjs/mocha/model_tests/importing.test.js
@@ -369,7 +369,7 @@ describe('importing', function () {
       Factory.createMappedTag(application),
     ])
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await application.sync.sync({ awaitAll: true })
 
@@ -401,7 +401,7 @@ describe('importing', function () {
       Factory.createMappedTag(application),
     ])
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await Factory.safeDeinit(application)
     application = await Factory.createInitAppWithFakeCrypto()
@@ -429,7 +429,7 @@ describe('importing', function () {
       Factory.createMappedTag(application),
     ])
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await Factory.safeDeinit(application)
     application = await Factory.createInitAppWithFakeCrypto()
@@ -458,7 +458,7 @@ describe('importing', function () {
       text: 'On protocol version 003.',
     })
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await Factory.safeDeinit(application)
     application = await Factory.createInitAppWithFakeCrypto()
@@ -539,7 +539,7 @@ describe('importing', function () {
       text: 'On protocol version 004.',
     })
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await Factory.safeDeinit(application)
     application = await Factory.createInitAppWithFakeCrypto()
@@ -569,7 +569,7 @@ describe('importing', function () {
       text: 'On protocol version 004.',
     })
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await Factory.safeDeinit(application)
     application = await Factory.createInitAppWithFakeCrypto()
@@ -607,7 +607,7 @@ describe('importing', function () {
       text: 'On protocol version 003.',
     })
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await Factory.safeDeinit(application)
     application = await Factory.createInitAppWithFakeCrypto()
@@ -645,7 +645,7 @@ describe('importing', function () {
       text: 'On protocol version 004.',
     })
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await Factory.safeDeinit(application)
     application = await Factory.createInitAppWithFakeCrypto()
@@ -678,7 +678,7 @@ describe('importing', function () {
       text: 'On protocol version 004.',
     })
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
     delete backupData.keyParams
 
     await Factory.safeDeinit(application)
@@ -703,7 +703,7 @@ describe('importing', function () {
       text: 'On protocol version 004.',
     })
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
     backupData.items = backupData.items.filter(
       (payload) => payload.content_type !== ContentType.ItemsKey,
     )
@@ -864,7 +864,7 @@ describe('importing', function () {
 
     await application.sync.sync()
 
-    const backupData = await application.createEncryptedBackupFile()
+    const backupData = await application.createEncryptedBackupFileForAutomatedDesktopBackups()
 
     await Factory.safeDeinit(application)
     application = await Factory.createInitAppWithFakeCrypto()